Refactored the settings action
