+++ Resources/views/Backend/Setting/index.html.twig
@@ -1,0 +1,23 @@
+{% extends 'OpiferCmsBundle:Backend:base.html.twig' %}
+
+{% block body %}
+    <div class="page-header clearfix">
+        <h1 class="pull-left">
+            <span class="iconl icon-dashboard"></span>{{ 'setting.index.title'|trans|capitalize }}
+        </h1>
+    </div>
+
+    <div class="container">
+        <div class="row">
+            {{ form_start(form, {'style': 'horizontal'}) }}
+                {{ form_widget(form) }}
+
+                <div class="form-group form-footer">
+                    <div class="col-xs-12">
+                        <input type="submit" value="Save changes" class="btn btn-primary"/>
+                    </div>
+                </div>
+            {{ form_end(form) }}
+        </div>
+    </div>
+{% endblock %}
