--- Resources/views/Setting/index.html.twig
@@ -1,13 +1,0 @@
-{% extends 'OpiferCmsBundle::base.html.twig' %}
-
-{% block body %}
-    <div class="page-header clearfix">
-        <h1 class="pull-left">
-            <span class="iconl icon-dashboard"></span>{{ 'setting.index.title'|trans|capitalize }}
-        </h1>
-    </div>
-
-    <section class="page-section">
-        {{ form(form, {'style': 'horizontal'}) }}
-    </section>
-{% endblock %}
